+++ src/map/map.cpp
@@ -289,6 +289,7 @@ namespace Invader {
             this->tags.push_back(Tag(*this));
             auto &tag = this->tags[i];
             tag.p_tag_class_int = tags[i].primary_class;
+            tag.tag_data_index_offset = reinterpret_cast<const std::byte *>(tags + i) - this->tag_data;
 
             try {
                 auto *path = reinterpret_cast<const char *>(this->resolve_tag_data_pointer(tags[i].tag_path));